Allow running multiple lambdas
This modifies the behavior in the following way:

Lambdas defined in the `lambda` section in the configuration file get mounted
to `/api/:lambda_name`. Subpaths that are not assigned to a lambda are
dispatched to the `main` Lambda if defined. As a consequence defining `main` is
no longer mandatory.

For example

```
lambda:
    foo:
      handler: index.handler
      memory: 512
    main:
      handler: index.handler
      memory: 512
```

Creates a Lambda under `/api/foo`. All other paths are routed to the main
lambda.